+++ src/athreading/generator.py
@@ -19,13 +19,13 @@ else:  # pragma: not covered
 
 from athreading.aliases import AsyncGeneratorContext
 
-__all__ = ["generate", "ThreadedAsyncGenerator"]
+__all__ = ["ThreadedAsyncGenerator", "generate"]
 
 
-ParamsT = ParamSpec("ParamsT")
-YieldT = TypeVar("YieldT")
+_ParamsT = ParamSpec("_ParamsT")
+_YieldT = TypeVar("_YieldT")
 _YieldT_co = TypeVar("_YieldT_co", covariant=True)
-SendT = TypeVar("SendT")
+_SendT = TypeVar("_SendT")
 _SendT_co = TypeVar("_SendT_co", covariant=True)
 
 
@@ -33,32 +33,35 @@ _SendT_co = TypeVar("_SendT_co", covariant=True)
 def generate(
     fn: None = None,
     *,
+    buffer_maxsize: Optional[int] = None,
     executor: Optional[ThreadPoolExecutor] = None,
 ) -> Callable[
-    [Callable[ParamsT, Generator[_YieldT_co, _SendT_co, None]]],
-    Callable[ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
+    [Callable[_ParamsT, Generator[_YieldT_co, _SendT_co, None]]],
+    Callable[_ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
 ]:
     ...
 
 
 @overload
 def generate(
-    fn: Callable[ParamsT, Generator[_YieldT_co, _SendT_co, None]],
+    fn: Callable[_ParamsT, Generator[_YieldT_co, _SendT_co, None]],
     *,
+    buffer_maxsize: Optional[int] = None,
     executor: Optional[ThreadPoolExecutor] = None,
-) -> Callable[ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]]:
+) -> Callable[_ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]]:
     ...
 
 
 def generate(
-    fn: Optional[Callable[ParamsT, Generator[_YieldT_co, _SendT_co, None]]] = None,
+    fn: Optional[Callable[_ParamsT, Generator[_YieldT_co, _SendT_co, None]]] = None,
     *,
+    buffer_maxsize: Optional[int] = None,
     executor: Optional[ThreadPoolExecutor] = None,
 ) -> Union[
-    Callable[ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
+    Callable[_ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
     Callable[
-        [Callable[ParamsT, Generator[_YieldT_co, _SendT_co, None]]],
-        Callable[ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
+        [Callable[_ParamsT, Generator[_YieldT_co, _SendT_co, None]]],
+        Callable[_ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
     ],
 ]:
     """Decorates a thread-safe synchronous generator with a ThreadPoolExecutor and exposes a
@@ -66,72 +69,85 @@ def generate(
 
     Args:
         fn: Function returning a generator. Defaults to None.
-        executor (Optional[ThreadPoolExecutor], optional): Defaults to None.
+        buffer_maxsize: Maximum buffer size for background worker to pre-emptively pull
+            data into. Defaults to None (no priming).
+        executor: Defaults to None.
 
     Returns:
         Decorated generator function with lazy argument evaluation.
     """
     if fn is None:
-        return _create_generate_decorator(executor=executor)
+        return _create_generate_decorator(
+            buffer_maxsize=buffer_maxsize, executor=executor
+        )
 
     @functools.wraps(fn)
     def wrapper(
-        *args: ParamsT.args, **kwargs: ParamsT.kwargs
+        *args: _ParamsT.args, **kwargs: _ParamsT.kwargs
     ) -> AsyncGeneratorContext[_YieldT_co, _SendT_co]:
-        return ThreadedAsyncGenerator(fn(*args, **kwargs), executor=executor)
+        return ThreadedAsyncGenerator(
+            fn(*args, **kwargs), buffer_maxsize=buffer_maxsize, executor=executor
+        )
 
     return wrapper
 
 
 def _create_generate_decorator(
+    buffer_maxsize: Optional[int] = None,
     executor: Optional[ThreadPoolExecutor] = None,
 ) -> Callable[
-    [Callable[ParamsT, Generator[_YieldT_co, _SendT_co, None]]],
-    Callable[ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
+    [Callable[_ParamsT, Generator[_YieldT_co, _SendT_co, None]]],
+    Callable[_ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]],
 ]:
     def decorator(
-        fn: Callable[ParamsT, Generator[_YieldT_co, _SendT_co, None]],
-    ) -> Callable[ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]]:
+        fn: Callable[_ParamsT, Generator[_YieldT_co, _SendT_co, None]],
+    ) -> Callable[_ParamsT, AsyncGeneratorContext[_YieldT_co, _SendT_co]]:
         @functools.wraps(fn)
         def wrapper(
-            *args: ParamsT.args, **kwargs: ParamsT.kwargs
+            *args: _ParamsT.args, **kwargs: _ParamsT.kwargs
         ) -> AsyncGeneratorContext[_YieldT_co, _SendT_co]:
-            return ThreadedAsyncGenerator(fn(*args, **kwargs), executor=executor)
+            return ThreadedAsyncGenerator(fn(*args, **kwargs), buffer_maxsize, executor)
 
         return wrapper
 
     return decorator
 
 
-class ThreadedAsyncGenerator(AsyncGeneratorContext[YieldT, SendT]):
+class ThreadedAsyncGenerator(AsyncGeneratorContext[_YieldT, _SendT]):
     """Runs a thread-safe synchronous generator with a ThreadPoolExecutor and exposes a
     thread-safe AsyncGenerator.
     """
 
     def __init__(
         self,
-        generator: Generator[YieldT, SendT, None],
+        generator: Generator[_YieldT, _SendT, None],
+        buffer_maxsize: Optional[int] = None,
         executor: Optional[ThreadPoolExecutor] = None,
     ):
         """Initilizes a ThreadedAsyncGenerator from a synchronous generator.
 
         Args:
             generator: Synchronous generator.
+            buffer_maxsize: Maximum buffer size for background worker to pre-emptively pull
+                data into. Defaults to None (no priming).
             executor: Shared thread pool instance. Defaults to ThreadPoolExecutor().
         """
         self._yield_semaphore = asyncio.Semaphore(0)
         self._done_event = threading.Event()
-        self._send_queue: queue.Queue[Optional[SendT]] = queue.Queue()
-        self._yield_queue: queue.Queue[YieldT] = queue.Queue()
+        self._send_queue: queue.Queue[Optional[_SendT]] = queue.Queue()
+        self._yield_queue: queue.Queue[_YieldT] = queue.Queue(buffer_maxsize or 0)
+        if buffer_maxsize:
+            for _ in range(buffer_maxsize):
+                self._send_queue.put(None)
         self._generator = generator
         self._executor = executor
-        self._stream_future: Optional[asyncio.Future[None]] = None
+        self._worker_future: Optional[asyncio.Future[None]] = None
 
     @override
-    async def __aenter__(self) -> ThreadedAsyncGenerator[YieldT, SendT]:
+    async def __aenter__(self) -> ThreadedAsyncGenerator[_YieldT, _SendT]:
         self._loop = asyncio.get_running_loop()
-        self._stream_future = self._loop.run_in_executor(
-            self._executor, self.__stream_threadsafe
+        self._worker_future = self._loop.run_in_executor(
+            self._executor, self.__worker_threadsafe
         )
         return self
 
@@ -144,22 +160,22 @@ class ThreadedAsyncGenerator(AsyncGeneratorContext[YieldT, SendT]):
         /,
     ) -> None:
         # move to aclose
-        assert self._stream_future is not None
+        assert self._worker_future is not None
         self._done_event.set()
         self._yield_semaphore.release()
         self._send_queue.put(None)
-        await self._stream_future
+        await self._worker_future
 
     @override
-    async def __anext__(self) -> YieldT:
+    async def __anext__(self) -> _YieldT:
         assert (
-            self._stream_future is not None
+            self._worker_future is not None
         ), "Iteration started before entering context"
         self._send_queue.put(None)
         return await self.__get()
 
     @override
-    async def asend(self, value: Optional[SendT]) -> YieldT:
+    async def asend(self, value: Optional[_SendT]) -> _YieldT:
         """Send a value to the generator send queue"""
         self._send_queue.put(value)
         return await self.__get()
@@ -168,7 +184,7 @@ class ThreadedAsyncGenerator(AsyncGeneratorContext[YieldT, SendT]):
         """Closes the generator"""
         self._generator.close()
 
-    async def __get(self) -> YieldT:
+    async def __get(self) -> _YieldT:
         if not self._done_event.is_set() or not self._yield_queue.empty():
             await self._yield_semaphore.acquire()
             if not self._yield_queue.empty():
@@ -182,13 +198,13 @@ class ThreadedAsyncGenerator(AsyncGeneratorContext[YieldT, SendT]):
         __val: object = None,
         __tb: Optional[TracebackType] = None,
         /,
-    ) -> YieldT:
+    ) -> _YieldT:
         """Raise a custom exception immediately from the generator"""
         if isinstance(__typ, BaseException):
             raise __typ
         return self._generator.throw(__typ, __val, __tb)
 
-    def __stream_threadsafe(self) -> None:
+    def __worker_threadsafe(self) -> None:
         """Stream the synchronous itertor to the queue and notify the async thread."""
         try:
             while not self._done_event.is_set():