+++ tests/performance/test_call_benchmark.py
@@ -33,7 +33,7 @@ def square_athreaded(x: float, delay: float = 0.0):
 
 
 async def asquare_threaded(x: float, delay: float = 0.0):
-    return await asyncio.wrap_future(square_athreaded(x, delay))
+    return await asyncio.wrap_future(square_athreaded(x, delay))  # pyright: ignore
 
 
 @pytest.mark.benchmark(group="call", disable_gc=True, warmup=True)
@@ -41,18 +41,20 @@ def test_call_athreading_benchmark(benchmark):
     def test():
         return asyncio.run(asquare_athreading(2, 0.0))
 
-    assert 4 == benchmark(test)
+    assert benchmark(test) == 4
 
 
 @pytest.mark.benchmark(group="call", disable_gc=True, warmup=True)
 def test_call_threaded_benchmark(benchmark):
     def test():
         async def atest():
-            return await asyncio.wrap_future(square_athreaded(2, 0.0))
+            return await asyncio.wrap_future(
+                square_athreaded(2, 0.0)  # pyright: ignore
+            )
 
         return asyncio.run(atest())
 
-    assert 4 == benchmark(test)
+    assert benchmark(test) == 4
 
 
 @pytest.mark.benchmark(group="call", disable_gc=True, warmup=True)
@@ -64,7 +66,7 @@ def test_single_call_benchmark(benchmark: BenchmarkFixture, impl, delay):
     def test():
         return asyncio.run(impl(2, delay))
 
-    assert 4 == benchmark.pedantic(test)
+    assert benchmark.pedantic(test) == 4
 
 
 @pytest.mark.benchmark(group="call", disable_gc=True, warmup=True)
@@ -83,4 +85,4 @@ def test_multi_call_benchmark(benchmark, impl, num_tasks, delay):
 
     results = benchmark(test)
     for result in results:
-        assert 4 == result
+        assert result == 4